(PC-29248)[PRO] refactor: extract beamer logic into a hook
